+++ examples/test.py
@@ -56,13 +56,13 @@ grid, areas = sk.make_grid(R, nt=nt, nr=nr)
 S, H, H2 = sk.calculate_mels([('sss', '2s', '1s')], atom1, atom2,
                              R, grid, areas)
 
-print('B-H sps S  | %s' % check_abs(S[9], -0.34627316, eps))
-print('B-H sps H  | %s' % check_abs(H[9], 0.29069894, eps))
-print('B-H sps H2 | %s' % check_abs(H2[9], 0.29077536, eps))
+print('B-H sss S  | %s' % check_abs(S[-1], -0.34627316, eps))
+print('B-H sss H  | %s' % check_abs(H[-1], 0.29069894, eps))
+print('B-H sss H2 | %s' % check_abs(H2[-1], 0.29077536, eps))
 
 S, H, H2 = sk.calculate_mels([('sps', '1s', '2p')], atom2, atom1,
                              R, grid, areas)
 
-print('H-B sps S  | %s' % check_abs(S[8], -0.47147340, eps))
-print('H-B sps H  | %s' % check_abs(H[8], 0.33033262, eps))
-print('H-B sps H2 | %s' % check_abs(H2[8], 0.33028882, eps))
+print('H-B sps S  | %s' % check_abs(S[-2], -0.47147340, eps))
+print('H-B sps H  | %s' % check_abs(H[-2], 0.33033262, eps))
+print('H-B sps H2 | %s' % check_abs(H2[-2], 0.33028882, eps))